refactor(racingcar): Move input parsing logic new class
Moved input parsing logic for both inputs to new class
